multiple of two positive numbers

+++ 0x0C-more_malloc_free/101-mul.c
@@ -9,7 +9,7 @@
  */
 int main(int argc, char *argv[])
 {
-	long mul = 1;
+	unsigned long long mul = 1;
 	int i = 1, j;
 
 	if (argc != 3)
